move location types into api modules
