+++ common/pom.xml
@@ -21,6 +21,11 @@
         </dependency>
         <dependency>
             <groupId>org.example.age</groupId>
+            <artifactId>avs-api</artifactId>
+            <version>${revision}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.example.age</groupId>
             <artifactId>common-api</artifactId>
             <version>${revision}</version>
         </dependency>
@@ -37,6 +42,11 @@
         <dependency>
             <groupId>org.example.age</groupId>
             <artifactId>infra-api</artifactId>
+            <version>${revision}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.example.age</groupId>
+            <artifactId>site-api</artifactId>
             <version>${revision}</version>
         </dependency>
         <dependency>
