+++ common/src/main/java/org/example/age/common/avs/config/RegisteredSiteConfig.java
@@ -1,5 +1,6 @@
 package org.example.age.common.avs.config;
 
+import org.example.age.avs.api.SiteLocation;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.AgeThresholds;
 import org.example.age.data.utils.DataStyle;
