+++ common/src/main/java/org/example/age/common/avs/verification/internal/VerificationManagerImpl.java
@@ -9,10 +9,10 @@ import javax.inject.Named;
 import javax.inject.Provider;
 import javax.inject.Singleton;
 import org.example.age.api.HttpOptional;
+import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.api.data.AuthMatchData;
 import org.example.age.common.api.data.AuthMatchDataExtractor;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
-import org.example.age.common.avs.config.SiteLocation;
 import org.example.age.common.avs.store.RegisteredSiteConfigStore;
 import org.example.age.common.avs.store.VerifiedUserStore;
 import org.example.age.common.base.store.PendingStore;
