+++ common/src/test/java/org/example/age/common/avs/store/InMemoryRegisteredSiteConfigStoreTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import dagger.Component;
 import javax.inject.Singleton;
+import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
-import org.example.age.common.avs.config.SiteLocation;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.AgeThresholds;
 import org.junit.jupiter.api.BeforeEach;
