+++ common/src/test/java/org/example/age/common/avs/verification/VerificationManagerTest.java
@@ -14,8 +14,8 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 import org.assertj.core.data.Offset;
 import org.example.age.api.HttpOptional;
+import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.avs.config.RegisteredSiteConfig;
-import org.example.age.common.avs.config.SiteLocation;
 import org.example.age.common.avs.store.InMemoryRegisteredSiteConfigStoreModule;
 import org.example.age.common.avs.store.InMemoryVerifiedUserStoreModule;
 import org.example.age.common.avs.store.RegisteredSiteConfigStore;
