+++ site-api/src/main/java/module-info.java
@@ -1,10 +1,13 @@
 module org.example.age.site.api {
     exports org.example.age.site.api;
 
+    requires static org.immutables.value.annotations;
     requires com.fasterxml.jackson.databind;
+    requires com.google.common;
     requires dagger;
     requires java.compiler;
     requires javax.inject;
+    requires okhttp3;
     requires org.example.age.api;
     requires org.example.age.common.api;
     requires org.example.age.data;
