+++ site-service/src/main/java/org/example/age/site/service/config/SiteConfig.java
@@ -4,6 +4,7 @@ import java.security.PublicKey;
 import java.time.Duration;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.utils.DataStyle;
+import org.example.age.site.api.AvsLocation;
 import org.immutables.value.Value;
 
 /** Configuration for a site. */
